# eyeguard/OpenAPI.yaml
...
openapi: 3.0.3
info:
  title: EyeGuard API
  version: 1.0.0
  description: |
    Software-only simulation REST API for the EyeGuard SOC training platform.
servers:
  - url: https://localhost:8000
    description: Local FastAPI dev server
paths:
...
  /api/v1/devices:
    get:
      summary: List devices
      operationId: listDevices
      responses:
        '200':
          description: Devices collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    post:
      summary: Create simulated device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          description: Duplicate IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/devices/{deviceId}:
    get:
      summary: Retrieve device detail
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update device metadata
      operationId: updateDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
...
  /api/v1/simulation/devices:
    post:
      summary: Add simulated device to session
      operationId: addSimulationDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationDeviceCreate'
      responses:
        '201':
          description: Simulation device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationDevice'
  /api/v1/simulation/terminal:
    post:
      summary: Execute terminal command
      operationId: executeSimulationCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalCommandRequest'
      responses:
        '200':
          description: Command execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalCommandResponse'
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ip_address:
          type: string
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
        status:
          type: string
        last_seen_at:
          type: string
          format: date-time
    DeviceCreate:
      type: object
      required: [ip_address, hostname, device_type, owner_role]
      properties:
        ip_address:
          type: string
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
    DeviceUpdate:
      type: object
      properties:
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
        status:
          type: string
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        detected_at:
          type: string
          format: date-time
        source_ip:
          type: string
        destination_ip:
          type: string
        category:
          type: string
        severity:
          type: string
        status:
          type: string
        rationale:
          type: string
        action_taken:
          type: string
...
    SimulationDeviceCreate:
      type: object
      required: [ip_address, hostname, traffic_gb]
      properties:
        ip_address:
          type: string
        hostname:
          type: string
        traffic_gb:
          type: number
        device_type:
          type: string
    SimulationDevice:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        device:
          $ref: '#/components/schemas/Device'
    TerminalCommandRequest:
      type: object
      required: [session_id, command]
      properties:
        session_id:
          type: string
          format: uuid
        command:
          type: string
    TerminalCommandResponse:
      type: object
      properties:
        output:
          type: string
        alerts_triggered:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
...